feat: stages & matches (#4)
* add stage control with stage item and stage item input type

* set up types for everything else

* round and match generation

* teams and matches

* stageItem done

* players

* player router

* club done

* fixed up user, club, tournament, team

+++ controllers/club.controller.js
@@ -1,9 +1,18 @@
 import { ObjectId } from "mongodb";
 import z, { ZodError } from "zod";
-import { clubMembers, clubs, users } from "../config/db.js";
+import {
+    clubMembers,
+    clubs,
+    matches,
+    players,
+    stages,
+    teams,
+    tournaments,
+    users,
+} from "../config/db.js";
 
 const CREATE_SCHEMA = z.object({
-    name: z.string().trim().min(3).max(256), // todo: validate the format for special characters (dont allow)
+    name: z.string().trim().min(3).max(256),
     handle: z.string().trim().min(3).max(64),
 }).strict();
 
@@ -64,6 +73,11 @@ export async function getClub(req, res) {
         const club = await clubs.findOne({
             _id: new ObjectId(req.params.clubId),
         });
+
+        if (club == null) {
+            return res.status(404).json({ message: "Club not found" });
+        }
+
         return res.status(200).json({
             _id: club._id,
             name: club.name,
@@ -80,12 +94,23 @@ export async function getClub(req, res) {
     }
 }
 
-// todo: update / patch club
-
 /** @type {import("express").RequestHandler<{ clubId: string }>} */
 export async function deleteClub(req, res) {
     try {
         const clubId = new ObjectId(req.params.clubId);
+        const membership = await clubMembers.findOne({
+            clubId,
+            userId: new ObjectId(req.user.id),
+            role: "owner",
+        });
+
+        if (membership == null) {
+            return res.status(403).json({
+                message:
+                    "You don't have permission to delete teams in this tournament",
+            });
+        }
+
         const { deletedCount } = await clubs.deleteOne({ _id: clubId });
 
         if (deletedCount === 0) {
@@ -93,7 +118,19 @@ export async function deleteClub(req, res) {
                 .json({ message: "Could not find the specified club" });
         }
 
+        const tournamentsToDelete = await tournaments.find({ clubId })
+            .toArray();
+        const tournamentIds = tournamentsToDelete.map(t => t._id);
+
         await clubMembers.deleteMany({ clubId: clubId });
+
+        if (tournamentIds.length > 0) {
+            await tournaments.deleteMany({ clubId });
+            await teams.deleteMany({ tournamentId: { $in: tournamentIds } });
+            await players.deleteMany({ tournamentId: { $in: tournamentIds } });
+            await stages.deleteMany({ tournamentId: { $in: tournamentIds } });
+            await matches.deleteMany({ tournamentId: { $in: tournamentIds } });
+        }
 
         res.status(200)
             .json({ message: "Deleted club successfully" });
@@ -130,6 +167,7 @@ export async function getMyClubMemberships(req, res) {
                 },
             },
         ]).toArray();
+        console.log(clubMemberships);
         res.status(200).json(
             clubMemberships.map(
                 (
@@ -219,7 +257,7 @@ export async function getClubMembers(req, res) {
 }
 
 const ADD_MEMBER_SCHEMA = z.object({
-    query: z.string().trim(), // todo: validate the format for special characters (dont allow)
+    query: z.string().trim().regex(/^[a-zA-Z0-9_]+$/),
     role: z.enum(["member", "admin"]),
 }).strict();
 
@@ -280,6 +318,37 @@ export async function addClubMember(req, res) {
         });
     } catch (error) {
         console.error("Error during getting clubs:", error);
+        return res.status(500).json({ message: "Internal server error" });
+    }
+}
+
+/** @type {import("express").RequestHandler<{ clubId: string }>} */
+export async function getClubTournaments(req, res) {
+    try {
+        const membership = await clubMembers.findOne({
+            clubId: new ObjectId(req.params.clubId),
+            userId: new ObjectId(req.user.id),
+        });
+
+        if (membership == null) {
+            return res.status(403).json({
+                message: "You are not a member of this club",
+            });
+        }
+
+        const clubTournaments = await tournaments.find({
+            clubId: new ObjectId(req.params.clubId),
+        }).sort({ createdAt: -1 }).toArray();
+
+        if (clubTournaments.length === 0) {
+            return res.status(200).json({
+                message: "No tournaments found",
+            });
+        }
+        
+        res.status(200).json(clubTournaments);
+    } catch (error) {
+        console.error("Error during getting club tournaments:", error);
         return res.status(500).json({ message: "Internal server error" });
     }
 }
